fix another thing

+++ scripts/deployed/setup-swap.js
@@ -1,13 +1,20 @@
 #!/usr/bin/env node
 
 import { $ } from 'execa'
-import { writeFile } from 'node:fs/promises'
+import { writeFile, stat } from 'node:fs/promises'
+
+const swapExists = await stat('/swapfile').catch(() => false)
 
 console.log('setting up swapfile...')
-await $`fallocate -l 2048M /swapfile`
-await $`chmod 0600 /swapfile`
-await $`mkswap /swapfile`
-await writeFile('/proc/sys/vm/swappiness', '10')
-await $`swapon /swapfile`
-await writeFile('/proc/sys/vm/overcommit_memory', '1')
-console.log('swapfile setup complete')
+
+if (swapExists) {
+	console.log('swapfile already exists')
+} else {
+	await $`fallocate -l 2048M /swapfile`
+	await $`chmod 0600 /swapfile`
+	await $`mkswap /swapfile`
+	await writeFile('/proc/sys/vm/swappiness', '10')
+	await $`swapon /swapfile`
+	await writeFile('/proc/sys/vm/overcommit_memory', '1')
+	console.log('swapfile setup complete')
+}
